Adjust in delete contact - remove lock on delete

+++ Contacts37.Application/Usecases/Contacts/Commands/Delete/DeleteContactCommandHandler.cs
@@ -1,30 +1,28 @@
-﻿using AutoMapper;
-using Contacts37.Application.Common.Exceptions;
-using Contacts37.Application.Contracts.Persistence;
-using Contacts37.Domain.Entities;
+﻿using Contacts37.Application.Contracts.Persistence;
+using Contacts37.Domain.Exceptions;
 using MediatR;
 
 namespace Contacts37.Application.Usecases.Contacts.Commands.Delete
 {
-    public class DeleteContactCommandHandler : IRequestHandler<DeleteContactCommand, Unit>
+	public class DeleteContactCommandHandler : IRequestHandler<DeleteContactCommand, Unit>
     {
         private readonly IContactRepository _contactRepository;
-        private readonly IMapper _mapper;
 
-        public DeleteContactCommandHandler(IContactRepository contactRepository,
-            IMapper mapper)
+        public DeleteContactCommandHandler(IContactRepository contactRepository)
         {
             _contactRepository = contactRepository ?? throw new ArgumentNullException(nameof(contactRepository));
-            _mapper = mapper ?? throw new ArgumentNullException(nameof(mapper));
         }
 
         public async Task<Unit> Handle(DeleteContactCommand command, CancellationToken cancellationToken)
-        {
-            var contact = _mapper.Map<Contact>(command);
+		{
+            var contact = await _contactRepository.GetAsync(command.Id);
 
-            await _contactRepository.DeleteAsync(contact);
+            if (contact == null)
+				throw new InvalidIdException(command.Id);
 
-            return new Unit();
+			await _contactRepository.DeleteAsync(contact);
+
+            return Unit.Value;
         }
-    }
+	}
 }
